Add support for tracing the path by following the differential equation.
It turns out that this algorithm looks suspiciously like, possibly
equivalent to the extrapolate-and-nearest-point algorithm, although
I'm not sure I can be quite bothered to prove this.

However, the way it's written right now is not stable, as I think my
quick-and-dirty solver makes assumptions about suitably small
derivatives that we get convergence rather than oscillation, which is
not the case in reality. Should be fun to fix...